Move all notification views to ggrc module
