+++ src/ggrc_workflows/views/__init__.py
@@ -3,17 +3,13 @@
 # Created By: dan@reciprocitylabs.com
 # Maintained By: urban@reciprocitylabs.com
 
-from datetime import date, datetime
+from datetime import date
 from flask import render_template, redirect, url_for
-from ggrc.rbac import permissions
 from jinja2 import Environment, PackageLoader
-from werkzeug.exceptions import Forbidden
 
 from ggrc import db
-from ggrc import notifications
 from ggrc.app import app
 from ggrc.login import login_required
-from ggrc.notifications import common
 from ggrc.views.cron import run_job
 from ggrc_workflows import start_recurring_cycles
 from ggrc_workflows.models import Workflow
@@ -22,28 +18,6 @@ env = Environment(loader=PackageLoader('ggrc', 'templates'))
 
 # TODO: move these views to ggrc_views and all the functions to notifications
 # module.
-
-
-def show_pending_notifications():
-  if not permissions.is_admin():
-    raise Forbidden()
-  _, notif_data = notifications.get_pending_notifications()
-
-  for day, day_notif in notif_data.iteritems():
-    for user_email, data in day_notif.iteritems():
-      data = common.modify_data(data)
-  pending = env.get_template("notifications/view_pending_digest.html")
-  return pending.render(data=sorted(notif_data.iteritems()))
-
-
-def show_todays_digest_notifications():
-  if not permissions.is_admin():
-    raise Forbidden()
-  _, notif_data = notifications.get_todays_notifications()
-  for user_email, data in notif_data.iteritems():
-    data = common.modify_data(data)
-  todays = env.get_template("notifications/view_todays_digest.html")
-  return todays.render(data=notif_data)
 
 
 def _get_unstarted_workflows():
@@ -89,14 +63,6 @@ def start_unstarted_cycles():
 
 
 def init_extra_views(app):
-  app.add_url_rule(
-      "/_notifications/show_pending", "show_pending_notifications",
-      view_func=login_required(show_pending_notifications))
-
-  app.add_url_rule(
-      "/_notifications/show_todays_digest", "show_todays_digest_notifications",
-      view_func=login_required(show_todays_digest_notifications))
-
   app.add_url_rule(
       "/admin/unstarted_cycles",
       view_func=login_required(unstarted_cycles))
